requirements.txt

fastapi
openai
transformers
torch
faiss-cpu
shap
uvicorn
